Move Input and PermutationIter to own file
