+++ src/lib.rs
@@ -1,4 +1,6 @@
+pub mod input;
 pub mod parser;
+pub mod permutation_iter;
 pub mod sanitize;
 
 use anyhow::{anyhow, Context, Result};
